link-service/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.2
psycopg2-binary==2.9.11
requests==2.32.5

